pipeline/requirements.txt
tensorflow==2.11.0
pandas
numpy
requests
psycopg2-binary
